+++ pyproject.toml
@@ -62,6 +62,7 @@ tests = [
     "coverage[toml]",
     "requests",
     "httpx",
+    "httpx2",
 ]
 docs = [
     "sphinx>=8",
@@ -86,6 +87,7 @@ typing = [
     "types-defusedxml",
     "types-tqdm",
     "httpx",
+    "httpx2",
 ]
 docs-lint = [
     { include-group = "docs" },
